+++ source/constructs/bin/live-streaming.ts
@@ -15,6 +15,10 @@
 import 'source-map-support/register';
 import * as cdk from '@aws-cdk/core';
 import { LiveStreaming } from '../lib/live-streaming';
+import { AwsSolutionsChecks } from 'cdk-nag';
 
 const app = new cdk.App();
-new LiveStreaming(app, 'LiveStreaming');
+new LiveStreaming(app, 'LiveStreaming'); // NOSONAR
+
+ //cdk nag
+ cdk.Aspects.of(app).add(new AwsSolutionsChecks({ verbose: true }));